feat #14: 이슈 목록 조회 API 구현
- 향후 리팩토링 예정
- 페이징, 필터 기능 미구현

+++ be/src/main/java/com/issuetracker/dto/IssueListDto.java
@@ -16,8 +16,8 @@ public class IssueListDto {
     private List<FilterUserDto> userList;
     private List<FilterLabelDto> labelList;
     private List<FilterMilestoneDto> milestoneList;
-    private int countAllLabels;
-    private int countAllMilestones;
-    private int countOpenedIssues;
-    private int countClosedIssues;
+    private Integer countAllLabels;
+    private Integer countAllMilestones;
+    private Long countOpenedIssues;
+    private Long countClosedIssues;
 }
